Updated to MC 1.8.0 release.

+++ src/MiNET/MiNET/Entities/Hostile/Guardian.cs
@@ -29,7 +29,7 @@ namespace MiNET.Entities.Hostile
 {
 	public class Guardian : HostileMob, IAgeable
 	{
-		public Guardian(Level level) : base((int) EntityType.Guardian, level)
+		public Guardian(Level level) : base(EntityType.Guardian, level)
 		{
 			Width = Length = 0.7;
 			Height = 2.4;
