+++ src/MiNET/MiNET/Entities/Hostile/Husk.cs
@@ -29,7 +29,7 @@ namespace MiNET.Entities.Hostile
 {
 	public class Husk : HostileMob, IAgeable
 	{
-		public Husk(Level level) : base((int) EntityType.Husk, level)
+		public Husk(Level level) : base(EntityType.Husk, level)
 		{
 			Width = Length = 0.6;
 			Height = 1.8;
