+++ src/MiNET/MiNET/Entities/Hostile/Silverfish.cs
@@ -29,7 +29,7 @@ namespace MiNET.Entities.Hostile
 {
 	public class Silverfish : HostileMob
 	{
-		public Silverfish(Level level) : base((int) EntityType.Silverfish, level)
+		public Silverfish(Level level) : base(EntityType.Silverfish, level)
 		{
 			Width = Length = 0.4;
 			Height = 0.3;
