+++ src/MiNET/MiNET/Entities/Hostile/ZombiePigman.cs
@@ -29,7 +29,7 @@ namespace MiNET.Entities.Hostile
 {
 	public class ZombiePigman : HostileMob
 	{
-		public ZombiePigman(Level level) : base((int) EntityType.ZombiePigman, level)
+		public ZombiePigman(Level level) : base(EntityType.ZombiePigman, level)
 		{
 			Width = Length = 0.6;
 			Height = 1.8;
