+++ src/MiNET/MiNET/Entities/Hostile/ZombieVillager.cs
@@ -29,7 +29,7 @@ namespace MiNET.Entities.Hostile
 {
 	public class ZombieVillager : HostileMob
 	{
-		public ZombieVillager(Level level) : base((int) EntityType.ZombieVillager, level)
+		public ZombieVillager(Level level) : base(EntityType.ZombieVillager, level)
 		{
 			Width = Length = 0.6;
 			Height = 1.8;
